+++ payments/RELEASES.md
@@ -128,3 +128,13 @@ Based on:
 - [typescript v2.4.1] payments
 ### Releases
 - [NPM v2.4.1] https://www.npmjs.com/package/@wingspan/payments/v/2.4.1 - payments
+
+## 2024-09-09 01:31:53
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0-oas3.1 ./openapi/payment_speakeasy_edit.yaml
+- Speakeasy CLI 1.392.1 (2.411.9) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v2.5.0] payments
+### Releases
+- [NPM v2.5.0] https://www.npmjs.com/package/@wingspan/payments/v/2.5.0 - payments